Revert "Each spatial layer can only have 1 QP value."
This reverts commit 962b3935e44053641764beb0bd095540fe8cbd64.

Reason for revert: Breaks downstream tests.

Original change's description:
> Each spatial layer can only have 1 QP value.
>
> As the code is now, it looks like it accepts that a spatial layer can have more than 1 QP value. These QP values according to the code are summed. However, to my best knowledge this cannot be the case and makes the code hard to read. Therefore, updating it with a check such that it checks that each spatial layer only have 1 QP value and would be easier for a future code reader.
>
> Bug: webrtc:357636606
> Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/358101
> Commit-Queue: Emil Vardar (xWF) <vardar@google.com>
> Cr-Commit-Position: refs/heads/main@{#42736}

Bug: webrtc:357636606
No-Presubmit: true
No-Tree-Checks: true
No-Try: true
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/358721
Bot-Commit: rubber-stamper@appspot.gserviceaccount.com <rubber-stamper@appspot.gserviceaccount.com>
Owners-Override: Mirko Bonadei <mbonadei@webrtc.org>
Commit-Queue: Mirko Bonadei <mbonadei@webrtc.org>
Cr-Commit-Position: refs/heads/main@{#42737}

+++ test/pc/e2e/analyzer/video/default_video_quality_analyzer.cc
@@ -365,8 +365,7 @@ void DefaultVideoQualityAnalyzer::OnFrameEncoded(
   // TODO(https://crbug.com/webrtc/14891): If we want to support a mix of
   // simulcast and SVC we'll also need to consider the case where we have both
   // simulcast and spatial indices.
-  // In singlecast, spatial/simulcast is expected not to be set.
-  int stream_index = encoded_image.SpatialIndex().value_or(
+  size_t stream_index = encoded_image.SpatialIndex().value_or(
       encoded_image.SimulcastIndex().value_or(0));
   frame_in_flight.OnFrameEncoded(
       now, time_between_encoded_frames, encoded_image._frameType,
